+++ pymhf/utils/winapi.py
@@ -2,8 +2,10 @@
 
 import ctypes
 from ctypes import wintypes
+from typing import Protocol
 
 import pymem
+import pymem.ressources.structure
 
 GetModuleFileNameExA = ctypes.windll.psapi.GetModuleFileNameExA
 GetModuleFileNameExA.restype = wintypes.DWORD
@@ -38,6 +40,24 @@ SetLayeredWindowAttributes.argtypes = [
     wintypes.DWORD,
 ]
 
+
+SetLastError = ctypes.windll.kernel32.SetLastError
+SetLastError.argtypes = [wintypes.DWORD]
+
+
+GetLastError = ctypes.windll.kernel32.GetLastError
+GetLastError.restype = wintypes.DWORD
+
+
+VirtualQuery = ctypes.windll.kernel32.VirtualQuery
+VirtualQuery.argtypes = [
+    wintypes.LPCVOID,
+    ctypes.c_void_p,
+    ctypes.c_size_t,
+]
+VirtualQuery.restype = ctypes.c_size_t
+
+
 MAX_EXE_NAME_SIZE = 1024
 WS_EX_LAYERED = 0x00080000  # layered window
 GWL_EXSTYLE = -20  # "extended window style"
@@ -57,3 +77,22 @@ def set_window_transparency(hwnd: int, alpha: float):
     SetWindowLongA(hwnd, GWL_EXSTYLE, GetWindowLongA(hwnd, GWL_EXSTYLE) | WS_EX_LAYERED)
     rgb = wintypes.RGB(10, 10, 10)
     SetLayeredWindowAttributes(hwnd, rgb, int(255 * alpha), LWA_ALPHA)
+
+
+class MemoryInfo(Protocol):
+    BaseAddress: int
+    AllocationBase: int
+    AllocationProtect: int
+    RegionSize: int
+    State: int
+    Protect: int
+    Type: int
+
+
+def QueryAddress(address: int) -> MemoryInfo:
+    mbi = pymem.ressources.structure.MEMORY_BASIC_INFORMATION()
+    SetLastError(0)
+    VirtualQuery(address, ctypes.byref(mbi), ctypes.sizeof(mbi))
+    if errcode := GetLastError():
+        raise ValueError(f"There was an error accessing address 0x{address:X}: {errcode}")
+    return mbi